Work on WorldCoords<->Chunk Coords calculation

+++ include/Collision.h
@@ -22,8 +22,8 @@ struct Ray {
 	Ray() = default;
 	Ray(glm::vec3 start, glm::vec3 dir);
 
-	std::vector<glm::ivec3> getAffectedBlocks(float length) const;
-	glm::ivec3 getIntersectionFace(glm::ivec3 block) const;
+	std::vector<glm::vec3> getAffectedBlocks(float length) const;
+	glm::vec3 getIntersectionFace(glm::vec3 block) const;
 };
 
 typedef std::pair<glm::vec3, glm::vec3> Line;
